third_party/blink/web_tests/fast/dom/shadow/content-reprojection-fallback-crash-expected.txt

CONSOLE WARNING: line 13: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
A crash should not happen when fallback elements are reprojected.

PASS distributedNodes.item(0) is shadowRoot1.getElementById("fallback")
PASS distributedNodes.item(1) is shadowRoot1.getElementById("div")
PASS successfullyParsed is true

TEST COMPLETE
